requirements.txt
django
djangorestframework
django-ckeditor
googletrans==4.0.0-rc1
redis